Change state to contain list of entities
Instead of having a board like structure, and having to decide if it
should be a matrix or a hash indexed by position, I decided to go with a
flat list of entities. Since each entity has its own position, this does
not lose any information, and we can use whatever representation we want
when we need it. For example, I now include by_type and by_position
tables in the ToPlayer structure, so that implementers can use whatever
is most useful for them.